CHERI docker-release: docker/release files for CHERI releases of Zephyr
Added an orphan branch to contain CHERI specific information such
as release files and docker files associated with a particular
CHERI-Zephyr release.

The docker files will build and run the CHERI toolchain with
Zephyr. This additional information is not normally included
in the code base of the zephyr repository.

Added files for:
CHERI-RISCV64-V1.0
CHERI-RISCV64-V2.0

Signed-off-by: Jennifer Jackson <j.jackson@bham.ac.uk>