driver: modbus: CHERI: Added modbus demonstrator for CHERI QEMU
Add a copy of the modbus demonstrator specifically modified to run on
CODASIP CHERI CPUs under QEMU. Current version compiles with and
without purecap mode, but only runs reliably in non-purecap. To run
the demo, a patched QEMU with second UART1 is currently required.
This patch hence also adds a DTSI for this QEMU variant. This patch
also makes NS16550 driver IRQs compatible to CHERI builds.

Signed-off-by: David Oswald <david.f.oswald@durham.ac.uk>